pull out prompt helpers to separate file
