+++ services/QuillLMS/client/app/bundles/Staff/components/evidence/activitySessions/sessionOverview.tsx
@@ -4,7 +4,7 @@ import * as moment from 'moment'
 import PromptTable from './promptTable';
 
 import { DataTable, Error, Spinner } from '../../../../Shared/index';
-import { getPromptForActivitySession } from "../../../helpers/evidence";
+import { getPromptForActivitySession } from "../../../helpers/evidence/promptHelpers";
 import { BECAUSE, BUT, SO } from '../../../../../constants/evidence';
 
 const SessionOverview = ({ activity, rules, sessionData }) => {
